+++ src/structures/Command.js
@@ -5,8 +5,6 @@ const defaultOptions = {
   cooldown: 0,
   aliases: [],
   permLevel: 0,
-  description: undefined,
-  usage: undefined,
 };
 
 /**
@@ -38,10 +36,16 @@ module.exports = class Command extends Base {
     this.cooldown = this._options.cooldown;
     this.aliases = this._options.aliases;
     this.permLevel = this._options.permLevel;
-    this.description = this._options.description;
-    this.usage = this._options.usage;
     this.cooldowns = new Collection();
   }
 
   noPermsRun() {}
+
+  cdRun() {}
+
+  resetCooldowns(...ids) {
+    if (ids.length > 0) ids.forEach(id => this.cooldowns.delete(id));
+    else this.cooldowns.clear();
+    return this.cooldowns;
+  }
 };