+++ src/structures/MySql.js
@@ -1,19 +1,13 @@
 const { EventEmitter } = require('events');
 const mysql = require('mysql');
-const Model = require('../util/SqlModel');
+const SqlModel = require('../util/SqlModel');
 const Types = require('../util/Types');
+const Collection = require('../util/Collection');
 
-class MySql {
+module.exports = class MySql extends EventEmitter {
   constructor(url) {
-    // TODO: Options
-    /**
-     * @name MySql#_models
-     * @type {Array<Model>}
-     * @private
-     */
-    this._models = [];
+    this.collections = new Collection();
     this.url = url;
-    this.emitter = new EventEmitter();
     this.open(url);
   }
 
@@ -21,7 +15,7 @@ class MySql {
    * @returns {*}
    */
   close() {
-    this.emitter.emit('disconnected', this);
+    this.emit('disconnect');
     return this.db.end();
   }
 
@@ -32,18 +26,18 @@ class MySql {
    */
   open(url) {
     if (!url) url = this.url;
-    if (typeof url !== 'string') {
-      throw new TypeError('MySql uri must be a string.');
-    }
+    // if (typeof url !== 'string') {
+    //   throw new TypeError('MySql uri must be a string.');
+    // }
     this.url = url;
     this.db = mysql.createConnection(this.url);
     return new Promise((res, rej) => {
-      this.db.connect(err => {
+      this.db.connect((err) => {
         if (err) {
-          this.emitter.emit('error', err);
+          this.emit('error', err);
           return rej(err);
         }
-        this.emitter.emit('connected', this);
+        this.emit('connect');
         res(this.db);
       });
     });
@@ -62,7 +56,11 @@ class MySql {
     if (typeof name !== 'string') {
       throw new TypeError('Name argument must be a string.');
     }
-    if (this._models.find(e => e.toLowerCase() === name.toLowerCase())) {
+    if (
+      [...this.collections.keys()]
+        .map((k) => k.toLowerCase())
+        .includes(name.toLowerCase())
+    ) {
       throw new ReferenceError(`Model with name ${name} already exists`);
     }
     if ({}.hasOwnProperty.call(this, name)) {
@@ -72,28 +70,32 @@ class MySql {
       throw new TypeError('Options argument must be an object.');
     }
     const defaultOptions = {};
-    for (const key in options) {
-      if ({}.hasOwnProperty.call(options, key)) {
-        if (typeof options[key].type === 'function') {
-          options[key].type = options[key].type();
-        }
-        if (!options[key].type.db.includes('mysql')) {
-          throw new Error('Used no-sql data type for sql.');
-        }
-        defaultOptions[key] = options[key].default;
-        options[key] = options[key].type.mySqlType;
+    for (const key of Object.keys(options)) {
+      let type;
+      let defaults;
+      if (typeof options[key] === 'function') type = options[key];
+      if (typeof options[key].type === 'function') {
+        type = options[key].type;
+        defaults = options[key].default || undefined;
       }
+      type = type();
+      if (!type.db.includes('mysql')) {
+        throw new Error('No-sql data types are not allowed in sql.');
+      }
+      defaultOptions[key] = defaults;
+      options[key] = type.mySqlType;
     }
     options._id = 'VARCHAR(20)';
     defaultOptions._id = undefined;
-    this._models.push(name);
-    this[name] = new Model(
-      this.db,
-      name.toLowerCase(),
-      options,
-      defaultOptions
+    this.collections.set(
+      name,
+      new SqlModel(this.db, name.toLowerCase(), options, defaultOptions)
     );
     return this;
+  }
+
+  getCollection(name) {
+    return this.collections.get(name);
   }
 
   static get Types() {
@@ -104,6 +106,4 @@ class MySql {
     }
     return Types;
   }
-}
-
-module.exports = MySql;
+};